+++ src/main/java/net/minecraft/block/BlockLilyPad.java
@@ -60,7 +60,7 @@ public class BlockLilyPad extends BlockBush {
 		if (pos.getY() >= 0 && pos.getY() < 256) {
 			IBlockState iblockstate = worldIn.getBlockState(pos.down());
 			Material material = iblockstate.getMaterial();
-			return material == Material.WATER && iblockstate.getValue(BlockLiquid.LEVEL).intValue() == 0 || material == Material.ICE;
+			return material == Material.WATER && iblockstate.getValue(BlockLiquid.LEVEL) == 0 || material == Material.ICE;
 		} else {
 			return false;
 		}
