+++ src/main/java/net/minecraft/block/BlockPotato.java
@@ -39,7 +39,7 @@ public class BlockPotato extends BlockCrops {
 
 	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos) {
 
-		return POTATO_AABB[state.getValue(getAgeProperty()).intValue()];
+		return POTATO_AABB[state.getValue(getAgeProperty())];
 	}
 
 }
