+++ src/main/java/net/minecraft/block/BlockReed.java
@@ -27,7 +27,7 @@ public class BlockReed extends Block {
 	protected BlockReed() {
 
 		super(Material.PLANTS);
-		setDefaultState(blockState.getBaseState().withProperty(AGE, Integer.valueOf(0)));
+		setDefaultState(blockState.getBaseState().withProperty(AGE, 0));
 		setTickRandomly(true);
 	}
 
@@ -46,13 +46,13 @@ public class BlockReed extends Block {
 				}
 
 				if (i < 3) {
-					int j = state.getValue(AGE).intValue();
+					int j = state.getValue(AGE);
 
 					if (j == 15) {
 						worldIn.setBlockState(pos.up(), getDefaultState());
-						worldIn.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(0)), 4);
+						worldIn.setBlockState(pos, state.withProperty(AGE, 0), 4);
 					} else {
-						worldIn.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(j + 1)), 4);
+						worldIn.setBlockState(pos, state.withProperty(AGE, j + 1), 4);
 					}
 				}
 			}
@@ -157,7 +157,7 @@ public class BlockReed extends Block {
 	 */
 	public IBlockState getStateFromMeta(int meta) {
 
-		return getDefaultState().withProperty(AGE, Integer.valueOf(meta));
+		return getDefaultState().withProperty(AGE, meta);
 	}
 
 	/**
@@ -165,7 +165,7 @@ public class BlockReed extends Block {
 	 */
 	public int getMetaFromState(IBlockState state) {
 
-		return state.getValue(AGE).intValue();
+		return state.getValue(AGE);
 	}
 
 	protected BlockStateContainer createBlockState() {
